backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
feedparser==6.0.11
httpx==0.28.1
beautifulsoup4==4.12.3
pyyaml==6.0.2
apscheduler==3.10.4
